touch: switch to multi-touch protocol B
In protocol A, the getevent will show all the signals for all the touches. This method is useful for system debugging.
In protocol B, the getevent will only show the signals with differences. This improves the performance in event handling between the Linux and the Android and is the preferred method for release.

Change-Id: I58e1a6d95fdbf57debf3d0ee1d9a57469949dfec