touch: fix reported xmax and ymax values.
The right edge ABS_MT_POSITION_X can be reported as 0xEF(239) which is greater than the “getevent” ABS MAX for the X position.
Similarly, for bottom to top, the bottom edge Y position can be reported as 0xEF(239).  This is greater than the “getevent” ABS MAX of 238 for the Y position.

This is caused by the Driver when it sets the limits used by the X, Y max values.  The Driver is subtracting 1 from the device reported limits.

Change-Id: I20702c0b1818a905acc2989e87dc884edd14c546